refactor: extract shared browser utilities to reduce duplication
