+++ packages/wxt/src/core/runners/wxt-runner.ts
@@ -8,9 +8,11 @@ import {
   type KnownTarget,
   type Runner as WxtRunnerInstance,
 } from '@wxt-dev/runner';
-import * as fs from 'node:fs/promises';
-import { constants as fsConstants } from 'node:fs';
-import path from 'node:path';
+import {
+  resolveChromiumBinaryForRemoteDebuggingPipe,
+  resolveProfilePath,
+  sanitizePathForWslWithGui,
+} from './browser-utils';
 
 const KNOWN_TARGETS = new Set<string>(Object.keys(KNOWN_BROWSER_PATHS));
 function isKnownTarget(value: string): value is KnownTarget {
@@ -63,12 +65,14 @@ export function createWxtRunner(): ExtensionRunner {
         userConfig?.binaries?.[browser],
         `binaries.${browser}`,
         runningInWslWithGui,
+        '[runner] ',
       );
 
       const browserBinaryOverride = !isFirefoxFamilyTarget(browser)
         ? await resolveChromiumBinaryForRemoteDebuggingPipe({
             chromiumBinary: binaryFromConfig,
             runningInWslWithGui,
+            loggerPrefix: '[runner] ',
           })
         : binaryFromConfig;
 
@@ -84,6 +88,7 @@ export function createWxtRunner(): ExtensionRunner {
           userConfig?.firefoxProfile,
           'firefoxProfile',
           runningInWslWithGui,
+          '[runner] ',
         );
 
         const dataPersistence =
@@ -128,6 +133,7 @@ export function createWxtRunner(): ExtensionRunner {
           userConfig?.chromiumProfile,
           'chromiumProfile',
           runningInWslWithGui,
+          '[runner] ',
         );
 
         const dataPersistence =
@@ -166,90 +172,4 @@ export function createWxtRunner(): ExtensionRunner {
       runner = undefined;
     },
   };
-}
-
-function sanitizePathForWslWithGui(
-  value: string | undefined,
-  label: string,
-  runningInWslWithGui: boolean,
-): string | undefined {
-  if (!runningInWslWithGui || value == null) return value;
-  if (isWindowsPath(value)) {
-    wxt.logger.warn(
-      `[runner] Ignoring ${label}="${value}" in WSL with GUI. Windows paths/binaries are incompatible with CDP pipe extension install. Install a Linux browser in WSL and omit this setting.`,
-    );
-    return undefined;
-  }
-  return value;
-}
-
-function isWindowsPath(value: string): boolean {
-  // Windows drive path: C:\...
-  if (/^[a-zA-Z]:\\/.test(value)) return true;
-  // WSL mounted drive: /mnt/c/...
-  if (/^\/mnt\/[a-zA-Z]\//.test(value)) return true;
-  // UNC-ish
-  if (value.startsWith('\\\\')) return true;
-  return false;
-}
-
-async function resolveChromiumBinaryForRemoteDebuggingPipe({
-  chromiumBinary,
-  runningInWslWithGui,
-}: {
-  chromiumBinary: string | undefined;
-  runningInWslWithGui: boolean;
-}): Promise<string | undefined> {
-  if (!runningInWslWithGui) return chromiumBinary;
-
-  const googleChromeRealBinary = '/opt/google/chrome/chrome';
-  const hasRealGoogleChrome = await isExecutable(googleChromeRealBinary);
-
-  if (chromiumBinary == null) {
-    if (hasRealGoogleChrome) return googleChromeRealBinary;
-    return chromiumBinary;
-  }
-
-  if (hasRealGoogleChrome && looksLikeGoogleChromeWrapper(chromiumBinary)) {
-    wxt.logger.warn(
-      `[runner] Using "${googleChromeRealBinary}" instead of "${chromiumBinary}" in WSL with GUI to keep the CDP pipe open (avoids "Remote debugging pipe file descriptors are not open").`,
-    );
-    return googleChromeRealBinary;
-  }
-
-  if (looksLikeGoogleChromeWrapper(chromiumBinary)) {
-    const resolved = await fs
-      .realpath(chromiumBinary)
-      .catch(() => chromiumBinary);
-    const sibling = path.join(path.dirname(resolved), 'chrome');
-    if (await isExecutable(sibling)) return sibling;
-  }
-
-  return chromiumBinary;
-}
-
-async function isExecutable(filePath: string): Promise<boolean> {
-  try {
-    await fs.access(filePath, fsConstants.X_OK);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function looksLikeGoogleChromeWrapper(filePath: string): boolean {
-  const base = path.basename(filePath);
-  if (base === 'google-chrome') return true;
-  if (base === 'google-chrome-stable') return true;
-  if (base === 'google-chrome-beta') return true;
-  if (base === 'google-chrome-dev') return true;
-  if (base === 'google-chrome-unstable') return true;
-  if (filePath === '/opt/google/chrome/google-chrome') return true;
-  return false;
-}
-
-function resolveProfilePath(projectRoot: string, profileDir: string): string {
-  return path.isAbsolute(profileDir)
-    ? profileDir
-    : path.resolve(projectRoot, profileDir);
 }
